:tractor: refactor: refactor folder structures
