+++ src/components/HomePage.vue
@@ -1,8 +1,8 @@
 <template>
     <section>
         <page-title :pagename="'OVH Icon Database'"></page-title>
-        <icons-filter></icons-filter>
-        <icons></icons>
+        <icon-filter></icon-filter>
+        <icon-list></icon-list>
         <download-panel v-if="hasSelectedIcon"></download-panel>
     </section>
 </template>
@@ -11,8 +11,8 @@
 import { mapActions, mapGetters } from 'vuex'
 
 import PageTitle from '@/components/PageTitle/PageTitle'
-import IconsFilter from '@/components/IconsFilter/IconsFilter'
-import IconsList from '@/components/IconsList/IconsList'
+import IconFilter from '@/components/Icon/IconFilter/IconFilter'
+import IconList from '@/components/Icon/IconList/IconList'
 import DownloadPanel from '@/components/DownloadPanel/DownloadPanel'
 
 export default {
@@ -35,9 +35,9 @@ export default {
         'fetchIcons'
     ]),
     components: {
-      IconsList,
+      IconList,
       PageTitle,
-      IconsFilter,
+      IconFilter,
 		  DownloadPanel
     }
 }
