+++ src/components/Icon/IconFilter/IconFilter.scss
